rename product component to allProduct and move to component folder
